added translation backend
important: the current translation method dosent use the proper google translate (cloud translation) API with a key but instead just uses end points. this was based off the original vaani connect web app which just uses API endpoints instead of the actual key. works fine for testing right now but for production we should do the proper API implementation and add that to the costs (or not)